+++ src/renderer/router.tsx
@@ -15,7 +15,7 @@ const GeminiSettings = React.lazy(() => import('./pages/settings/GeminiSettings'
 const SudocodeModelSettings = React.lazy(() => import('./pages/settings/SudocodeModelSettings'));
 const SkillSettings = React.lazy(() => import('./pages/settings/SkillSettings'));
 const CopilotSettings = React.lazy(() => import('./pages/settings/CopilotSettings'));
-const RuntimeSettings = React.lazy(() => import('./pages/settings/RuntimeSettings'));
+const RuntimeSettings = React.lazy(() => import('./pages/settings/runtime'));
 const SystemSettings = React.lazy(() => import('./pages/settings/system'));
 const ToolsSettings = React.lazy(() => import('./pages/settings/tools'));
 const WebuiSettings = React.lazy(() => import('./pages/settings/WebuiSettings'));
